activity prediction with cam added
